+++ src/components/form/b-input/i18n/ru.js
@@ -1,0 +1,33 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+module.exports = {
+	'b-input': {
+		'The value is not a number': 'Значение не является числом',
+		'The value does not match with an unsigned integer type': 'Значение не является целочисленным типом без знака',
+		'The value does not match with an integer type': 'Значение не является целочисленным типом',
+		'The value does not match with an unsigned float type': 'Значение не является типом с плавающей точкой без знака',
+		'The decimal part must be {precision} digits': 'Десятичная часть должна содержать {precision} чисел',
+		'The decimal part must be no more than {precision} digits': 'Десятичная часть должна быть меньше, чем {precision} чисел',
+		'The value must be at least {min}': 'Значение должно быть больше чем {min}',
+		'The value must be no more than {max}}': 'Значение должно быть меньше чем {max}',
+		"The value can't be parsed as a date": 'Значение не является датой',
+		'Date value must be in the past': 'Дата должна быть в прошлом',
+		"Date value can't be in the past": 'Дата не может быть в прошлом',
+		'Date value must be in the future': 'Дата должна быть в будущем',
+		"Date value can't be in the future": 'Дата не может быть в будущем',
+		'Date value must be at least "{date}"': 'Дата не может быть раньше чем "{date}"',
+		'Date value must be no more than "{date}"': 'Дата не может быть позднее чем "{date}"',
+		'Invalid email format': 'Недопустимый формат электронной почты',
+		'Password must contain only allowed characters': 'Пароль должен содержать только разрешенные символы',
+		'Password length must be at least {min} characters': 'Длина пароля должна составлять не менее {min} символов',
+		'Password length must be no more than {max} characters': 'Длина пароля должна составлять не более {max} символов',
+		'The old and new passwords are the same': 'Старый и новый пароль одинаковы',
+		"The passwords don't match": 'Пароли не совпадают'
+	}
+};