+++ src/components/form/b-select-date/const.ts
@@ -9,6 +9,9 @@
 import { Cache } from 'components/super/i-block/i-block';
 import type { Item } from 'components/form/b-select/b-select';
 
+const
+	t = i18n('b-select-date');
+
 export const months = [
 	t`January`,
 	t`February`,