+++ src/components/super/i-block/CHANGELOG.md
@@ -19,6 +19,18 @@ Changelog
 
 * Added complete documentation for the module
 
+## v3.38.0 (2023-02-20)
+
+#### :bug: Bug Fix
+
+* Fixed use of `i18n` function in default prop values
+
+## v3.35.0 (2023-02-14)
+
+#### :boom: Breaking Change
+
+* Changed `i18n` function type from prop to getter
+
 ## v3.30.0 (2022-10-19)
 
 #### :rocket: New Feature